add custom config for sw6

+++ prepare.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
 
+set -x
 
 projectDir="$PWD/project/"
 
@@ -13,6 +14,10 @@ if [[ ! -f "$projectDir/.qodana-profile.xml" ]]; then
 fi
 
 if [[ ! -f "$projectDir/qodana.yaml" ]]; then
-  cp default/qodana.yaml "$projectDir/qodana.yaml"
+  if [[ -e "./default/${PROJECT_NAME}/qodana.yaml" ]]; then
+    cp "./default/${PROJECT_NAME}/qodana.yaml" "$projectDir/qodana.yaml"
+  else
+    cp default/qodana.yaml "$projectDir/qodana.yaml"
+  fi
   echo "Copied qodana.yaml"
 fi
